additions: Added date formatting

+++ app/src/main/java/com/simba/imgurapp/data/models/ImageItemDetails.kt
@@ -19,7 +19,7 @@ data class ImageItemDetails(
 
     @SerializedName("datetime")
     @Expose
-    val postedAt: Int? = null,
+    val postedAt: Long? = null,
 
     @SerializedName("images")
     @Expose
